+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 -   Support building non-client TypeScript projects
 
+### Fixed
+
+-   Respect `#!` lines when adding license header
+
 ## [0.1.3](https://github.com/atomist/atomist-sdm/compare/0.1.1...0.1.3) - 2018-07-05
 
 ### Added